conpaas-director/requirements.txt
flask-sqlalchemy
simplejson
pycurl
pyopenssl
m2crypto
